+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `import.file` is removed. Use `import.path` instead for files and directories.
 - `keycloak.migrationKey` is removed. Use `import.cache-key` instead.
 - `keycloak.realm` is removed. Use `import.login-realm` to define the realm to login.
-- If you have defined requiredActions in your realm configure, make sure you have defined all in your json files. All not defined
+- If you have defined requiredActions, components or subcomponents in your realm configure, make sure you have defined all in your json files. All not defined
   actions will removed now by keycloak-config-cli. See: [docs/MANAGED.md](docs/MANAGED.md)
 
 ### Added
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Throw errors on unknown properties in config files
 - Add, update and remove clientScopes
 - Removed required actions if they not defined in import json.
+- Removed components if they not defined in import json.
+- Removed subcomponents if they not defined in import json.
 - Contrib behaivor of purging ressource via `import.manage.<type>` property. See: [docs/MANAGED.md](docs/MANAGED.md)
 
 ### Changed